componentに分けたよ
